Managed to remove lag when alien explodes.
The problem with the aliens exploding was caused by the addition of a light to the scene. This might be solvable if other lights in the scene can be removed. Then use only a single light for when there are explosions positioned in the average of the two explostion centers.

Liberator updated to allow for omni-directional monopropellent.
Monopropellent updated to be moved so it comes from a port rather than the center of the spaceship (available in class not yet implemented by ships).
Plannet class updated to inherit from celestial and celestical class created.
: